+++ server/src/game_server/gameObjects/player.ts
@@ -1,32 +1,32 @@
 import { CircleHitbox2D, Client, NullVec2, Numeric, RectHitbox2D, v2, Vec2 } from "common/scripts/engine/mod.ts"
 import { ActionPacket, InputAction, InputActionType } from "common/scripts/packets/action_packet.ts"
-import { PlayerAnimation, PlayerData } from "common/scripts/others/objectsEncode.ts";
-import { ActionsType, GameConstants, GameOverPacket } from "common/scripts/others/constants.ts";
-import { GInventory,GunItem,LItem} from "../player/inventory.ts";
-import { DamageSplash, UpdatePacket } from "common/scripts/packets/update_packet.ts";
-import { DamageParams } from "../others/utils.ts";
-import { type Obstacle } from "./obstacle.ts";
-import { ActionsManager } from "common/scripts/engine/inventory.ts";
-import { DamageReason, GameItem, InventoryItemType } from "common/scripts/definitions/utils.ts";
-import { type EquipamentDef } from "common/scripts/definitions/items/equipaments.ts";
-import { DamageSourceDef, DamageSources, GameItems, Weapons } from "common/scripts/definitions/alldefs.ts";
-import { type PlayerModifiers } from "common/scripts/others/constants.ts";
-import { AccessoriesManager } from "../player/accesories.ts";
-import { ServerGameObject } from "../others/gameObject.ts";
-import { type Loot } from "./loot.ts";
-import { Ammos } from "common/scripts/definitions/items/ammo.ts";
-import { type Group, type Team } from "../others/teams.ts";
-import { KillFeedMessageType } from "common/scripts/packets/killfeed_packet.ts";
+import { PlayerAnimation, PlayerData } from "common/scripts/others/objectsEncode.ts"
+import { ActionsType, GameConstants, GameOverPacket } from "common/scripts/others/constants.ts"
+import { GInventory,GunItem,LItem} from "../player/inventory.ts"
+import { DamageSplash, UpdatePacket } from "common/scripts/packets/update_packet.ts"
+import { DamageParams } from "../others/utils.ts"
+import { type Obstacle } from "./obstacle.ts"
+import { ActionsManager } from "common/scripts/engine/inventory.ts"
+import { DamageReason, GameItem, InventoryItemType } from "common/scripts/definitions/utils.ts"
+import { type EquipamentDef } from "common/scripts/definitions/items/equipaments.ts"
+import { DamageSourceDef, DamageSources, GameItems, Weapons } from "common/scripts/definitions/alldefs.ts"
+import { type PlayerModifiers } from "common/scripts/others/constants.ts"
+import { AccessoriesManager } from "../player/accesories.ts"
+import { ServerGameObject } from "../others/gameObject.ts"
+import { type Loot } from "./loot.ts"
+import { Ammos } from "common/scripts/definitions/items/ammo.ts"
+import { type Group, type Team } from "../others/teams.ts"
+import { KillFeedMessageType } from "common/scripts/packets/killfeed_packet.ts"
 import {SkinDef, Skins} from "common/scripts/definitions/loadout/skins.ts"
-import { type VehicleSeat } from "./vehicle.ts";
-import { Floors, FloorType } from "common/scripts/others/terrain.ts";
-import { BoostDef, Boosts, BoostType } from "common/scripts/definitions/player/boosts.ts";
-import { EffectInstance, Effects, SideEffect, SideEffectType } from "common/scripts/definitions/player/effects.ts";
-import { BotAi } from "../player/simple_bot_ai.ts";
-import { EmoteDef, Emotes } from "common/scripts/definitions/loadout/emotes.ts";
-import { GunDef } from "common/scripts/definitions/items/guns.ts";
-import { ProjectileDef } from "common/scripts/definitions/objects/projectiles.ts";
-import { Explosions } from "common/scripts/definitions/objects/explosions.ts";
+import { type VehicleSeat } from "./vehicle.ts"
+import { Floors, FloorType } from "common/scripts/others/terrain.ts"
+import { BoostDef, Boosts, BoostType } from "common/scripts/definitions/player/boosts.ts"
+import { EffectInstance, Effects, SideEffect, SideEffectType } from "common/scripts/definitions/player/effects.ts"
+import { BotAi } from "../player/simple_bot_ai.ts"
+import { EmoteDef, Emotes } from "common/scripts/definitions/loadout/emotes.ts"
+import { GunDef } from "common/scripts/definitions/items/guns.ts"
+import { ProjectileDef } from "common/scripts/definitions/objects/projectiles.ts"
+import { Explosions } from "common/scripts/definitions/objects/explosions.ts"
 
 export class Player extends ServerGameObject{
     oldPosition:Vec2